AutoAdminSaas/requirements.txt
streamlit
openai
fpdf
python-dotenv
pandas
watchdog